% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearity.R
\docType{data}
\name{elu_f}
\alias{elu_f}
\alias{identity_f}
\alias{relu_f}
\alias{sigmoid_f}
\title{Activation functions}
\format{An object of class \code{activation_f} of length 3.}
\usage{
elu_f

identity_f

relu_f

sigmoid_f
}
\description{
Activation function ojbects (class \code{activation_f}) store three functions:
\itemize{
   \item{\code{f}: The activation function (nonlinearity) itself}
   \item{\code{grad}: The activation function's gradient with respect to
       \code{x}}
   \item{\code{initialize_final_biases}: A function used internally to
       initialize the biases of a network's output layer}
}
}
\keyword{datasets}

