% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearity.R
\docType{data}
\name{elu_activator}
\alias{elu_activator}
\alias{identity_activator}
\alias{relu_activator}
\alias{sigmoid_activator}
\title{Activator objects and activation functions}
\format{An object of class \code{activator} of length 3.}
\usage{
elu_activator

identity_activator

relu_activator

sigmoid_activator
}
\description{
\code{activator} objects store three functions relating to activation functions:
\itemize{
   \item{\code{f}: The activation function (nonlinearity) itself}
   \item{\code{grad}: The activation function's gradient with respect to
       \code{x}}
   \item{\code{initialize_activatorinal_biases}: A function used internally to
       initialize the biases of a network's output layer}
}
}
\keyword{datasets}

