% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustable.R
\name{adjustable}
\alias{adjustable}
\title{Flag a distribution parameter for optimization}
\usage{
adjustable(x)
}
\arguments{
\item{x}{An argument to a \code{\link{make_distribution}}}
}
\description{
Flag a distribution parameter for optimization
}
\examples{
# This distribution object is marked as having a fixed standard deviation
# and a mean that should be optimized
dist = make_distribution("NO", mu = adjustable(0), sigma = 1)
}

