% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{backprop}
\alias{backprop}
\title{Backprop: calculate network gradients using backpropagation}
\usage{
backprop(network, state, par, ...)
}
\arguments{
\item{network}{a \code{network} object, as created by \code{\link{mistnet}}}

\item{state}{if a \code{network_sate} object is provided, it can be used
instead of producing a new one with \code{\link{feedforward}}. This may
save some computation time.}

\item{par}{A list or vector containing the parameters. If not included,
\code{network$par_list} will be used}

\item{...}{(currently not used)}
}
\value{
a \code{vector} of gradients.
}
\description{
Backprop: calculate network gradients using backpropagation
}

