% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{BMN}
\alias{BMN}
\alias{dBMN}
\alias{rBMN}
\title{Approximate binary Markov network distribution}
\usage{
BMN(x, mu, alpha, beta)

dBMN(x, mu, alpha, beta)

rBMN(n, alpha, beta)
}
\arguments{
\item{x}{a binary matrix}

\item{mu}{a matrix of marginal probabilities with the same dimensions as
\code{x}; calculated internally from \code{alpha} and \code{beta}.}

\item{alpha}{a vector of unary potentials, whose length equals the number of
columns in \code{x}}

\item{beta}{a square matrix of binary potentials with zeros along the
diagonal. The dimensions must equal the number of columns in \code{x}.}

\item{n}{number of random samples in \code{rBMN}}
}
\description{
Ising model, Markov random field, conditional random field, pseudolikelihood
}
\references{
Holger Höfling and Robert Tibshirani. 2009. "Estimation of Spase Binary
   Pairwise Markov Networks using Pseudo-likelihoods". \emph{Journal of Machine
   Learning Research}.

Jason D Lee, Trevor J Hastie. 2012. "Learning the Structure of Mixed Graphical
   Models. \emph{Journal of Machine Learning Research}.
}

