% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_distribution.R
\docType{data}
\name{dldx}
\alias{dldx}
\title{Partial derivatives with respect to x}
\format{A list of functions, named according to their
\link[gamlss.dist]{gamlss.family} abbreviations (e.g. "NO" for the normal
distribution).  Additional functions can be added to the list by the user,
making it possible to use the corresponding distributions as priors.

Currently, the only distribution with built-in support is
"\link[gamlss.dist]{NO}", the normal distribution}
\usage{
dldx
}
\description{
These gradients are used for optimizing x with respect to the log-prior density
(as opposed to optimizing the distribution's parameters with respect to x, as with dldm
and the other derivatives available from \code{\link{error_distribution}}
objects). Mostly used internally by \code{\link{make_gamlss_distribution}}.
}
\examples{
# find the gradient with respect to x using dldx
exact_gradient = dldx$NO(x = 1, mu = 2, sigma = pi)

# Compare with the numerical gradient for the same distribution
numerical_gradient = numDeriv::grad(function(x)dnorm(x = x, mean = 2, sd = pi, log = TRUE), x = 1)

stopifnot(all.equal(exact_gradient, numerical_gradient))
}
\keyword{datasets}

