% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{draw_samples.distribution}
\alias{draw_samples.distribution}
\title{Sample random numbers from a probability distribution}
\usage{
\method{draw_samples}{distribution}(object, ...)
}
\arguments{
\item{object}{an \code{\link{distribution}} object}

\item{...}{additional arguments passed to \code{\link{get_values}}}
}
\description{
Sample random numbers from a probability distribution
}

