% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{draw_samples}
\alias{draw_samples}
\title{Draw random samples from an object}
\usage{
draw_samples(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Draw random samples from an object
}

