% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ENO}
\alias{ENO}
\alias{dENO}
\alias{rENO}
\title{Normal distribution with empirical mean and variance}
\usage{
ENO(by)

dENO(x, log, by, ...)

rENO(n, x, ...)
}
\arguments{
\item{by}{one of \code{"row"}, \code{"col"}, or \code{"mat"}. Should
the empirical moments be estimated for each row, each column, or for the
full matrix?}

\item{x}{vector of quantiles}

\item{log}{logical; if TRUE, probabilities p are instead given as log(p)}

\item{...}{additional arguments passed to or from other methods (currently
not used)}

\item{n}{number of samples}
}
\description{
This distribution produces a Gaussian distribution whose moments are
dynamically set to match \code{mean(x)} and \code{sd(x)}. It is used for
empirical-Bayesian descriptions of the distribution of \code{x} (similar to
the distribution of coefficients in a mixed effects model).
}

