% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedforward.R
\name{feedforward}
\alias{feedforward}
\title{Calculate network state from its coefficients (feedforward)}
\format{a \code{list} of \code{list}s.
\itemize{
   \item{\code{inputs}: A list of input matrices for each layer.  The first layer's input
       is concatenated from the network's \code{x} and \code{z} matrices, while
       subsequent inputs come from the previous layer's \code{output}}.
   \item{\code{pre_activations}: A list of pre_activation matrices for each layer,
       equal to the matrix product of the layer's \code{input} and its
       \code{weights}, plus its \code{biases}}.
   \item{\code{outputs}: A list of outputs for each layer, produced by applying the layer's
       activation function to the \code{pre_activations}}.
}}
\usage{
feedforward(network, ...)
}
\arguments{
\item{network}{a \code{network} object, as created by \code{\link{mistnet}}}

\item{...}{(currently not used)}
}
\description{
Calculate network state from its coefficients (feedforward)
}

