% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-class.R
\name{get_values}
\alias{get_values}
\title{Get parameter values from a distribution object}
\usage{
get_values(distribution, ...)
}
\arguments{
\item{distribution}{a distribution object}

\item{...}{additional arguments, which will override values contained within
the \code{distribution}}
}
\description{
Get parameter values from a distribution object
}

