% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-class.R
\name{grad}
\alias{grad}
\title{Calculate the gradient of a distribution}
\usage{
grad(distribution, name, ...)
}
\arguments{
\item{distribution}{An \code{\link{distribution}} object}

\item{name}{One of \code{"mu"}, \code{"sigma"}, \code{"nu"}, \code{"tau"}, or \code{"x"}}

\item{...}{additional arguments passed to \code{\link{get_values}}}
}
\description{
Calculate the gradient of a distribution
}

