% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflate}
\alias{inflate}
\alias{inflate.default}
\alias{inflate.inflatable}
\alias{make_inflatable}
\title{"inflate" a vector by repeating rows or columns}
\usage{
inflate(x)

\method{inflate}{default}(x)

\method{inflate}{inflatable}(x)

make_inflatable(x, by, n)
}
\arguments{
\item{x}{a vector}

\item{by}{one of \code{"row"} or \code{"col"}. Should elements be repeated
across rows or across columns?}

\item{n}{number of times to repeat each element}
}
\description{
A vector can be tagged with the attribute \code{inflatable} by
\code{make_inflatable}. This is useful when a vector of adjustable parameters
needs to be re-used across the rows or columns of a matrix (e.g. if each row
of an error distribution has its own standard deviation).
}
\details{
\code{inflatable} vectors can be \code{inflate}d to repeat their
elements across the rows or columns in a newly-formed matrix, but are
otherwise unchanged. Calling \code{inflate()} on a non-inflatable vector
will leave the vector unchanged. The derivatives of \code{inflatable} vectors
are automatically handled separatly by \code{\link{backprop}()}
}

