% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Describe a layer of a neural network}
\usage{
layer(activator, n_nodes, weight_prior, weights = NULL, biases = NULL, ...)
}
\arguments{
\item{activator}{an \code{\link{activator}} object, determining the nonlinear
function performed by this layer and its derivatives}

\item{n_nodes}{a whole number determining the size of the layer. The network's
last layer should have the same number of nodes as there are columns in
\code{y}.}

\item{weight_prior}{a \code{\link{distribution}} object that acts as a prior
distribution for the weight matrix.}

\item{weights}{an optional weight matrix; if omitted, the
\code{\link{mistnet}} function will randomly initialize the weights}

\item{biases}{an optional bias vector; if omitted, the
\code{\link{mistnet}} function will randomly initialize the biases}

\item{...}{Additional objects to include in the layer (not currently used
by the \code{\link{mistnet}} function)}
}
\value{
a list of class \code{layer} containing the objects above
}
\description{
Describe a layer of a neural network
}

