% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_density.R
\name{log_density.distribution}
\alias{log_density.distribution}
\title{Calculate the log probability of a distribution}
\usage{
\method{log_density}{distribution}(object, ...)
}
\arguments{
\item{object}{an \code{\link{distribution}} object}

\item{...}{additional arguments passed to \code{\link{get_values}}}
}
\description{
Calculate the log probability of a distribution
}

