% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_density.R
\name{log_density.mistnet_network}
\alias{log_density.mistnet_network}
\title{Calculate the log-likelihood of a network object}
\usage{
\method{log_density}{mistnet_network}(object, state, par, include_penalties,
  ...)
}
\arguments{
\item{object}{A \code{network} object}

\item{state}{An optional \code{network_state} object. If not included,
state will be calculated by \code{\link{feedforward}}.}

\item{par}{a vector of network parameters}

\item{include_penalties}{Should penalty terms (e.g. from prior distributions)
be included in the log-density?}

\item{...}{Additional arguments (currently not used.)}
}
\description{
Calculate the log-likelihood of a network object
}

