% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_density.R
\name{log_density}
\alias{log_density}
\title{Calculate the log probability density of an object}
\usage{
log_density(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional arguments to be passed to other methods}
}
\description{
Calculate the log probability density of an object
}
\seealso{
\code{\link{log_density.error_distribution}},
   \code{\link{log_density.network}}
}

