% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_distribution.R
\name{make_gamlss_distribution}
\alias{make_gamlss_distribution}
\title{Make an \code{error_distribution} from a gamlss distribution}
\usage{
make_gamlss_distribution(abbreviation, ...)
}
\arguments{
\item{abbreviation}{An abbreviation matching a \link[gamlss.dist]{gamlss.family},
e.g. "NO" for the normal distribution or "ZIP" for the zero-inflated
Poisson distribution}

\item{...}{Additional arguments, possibly including \code{bd} (binomial denominator),
\code{sigma} (scale), \code{nu} (shape), or \code{tau} (shape),
depending on the distribution}
}
\value{
An \code{error_distribution} object.
}
\description{
Make an \code{error_distribution} from a gamlss distribution
}

