% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mistnet2}
\alias{mistnet2}
\alias{mistnet2-package}
\title{Neural Networks with Latent Random Variables.}
\description{
Estimate the joint distribution of many random variables (conditional on
a set of observed inputs) without strong assumptions about the functional
form of the distribution or conditional relationships.
}
\seealso{
\code{\link{mistnet}}
}

