% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mistnet_fit}
\alias{mistnet_fit}
\title{Fit a mistnet model}
\usage{
mistnet_fit(network, mistnet_optimizer = mistnet_fit_optimx, ...)
}
\arguments{
\item{network}{A \code{network} object, as produced by the
\code{\link{mistnet}} function}

\item{mistnet_optimizer}{A \code{function} for maximizing the penalized
likelihood of a mistnet model. By default, this is \code{\link{mistnet_fit_optimx}}.}

\item{...}{Additional arguments passed to the \code{mistnet_optimizer}}
}
\value{
A \code{network} object, like the one produced by \code{\link{mistnet}},
   but with updated parameters and with an extra list containing the
   \code{optimization_results}
}
\description{
Fit a mistnet model
}

