% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mistnet_fit_optimx}
\alias{mistnet_fit_optimx}
\title{Optimize a mistnet model using the \code{\link[optimx]{optimx}} package}
\usage{
mistnet_fit_optimx(network, fn, gr, method = "L-BFGS-B", itnmax = 1000,
  control = list(maximize = TRUE, starttests = TRUE), hessian = FALSE)
}
\arguments{
\item{network}{A \code{network} object, as produced by the
\code{\link{mistnet}} function}

\item{fn}{A function of \code{par} that returns a scalar describing the
(penalized) log-likelihood}

\item{gr}{A function of \code{par} that returns a vector containing the
gradient of the (penalized) log-likelihood with respect to all the
parameters in \code{par}.}

\item{method}{The name of one \code{method} used by \code{\link[optimx]{optimx}}.
By default, this is "L-BFGS-B", which uses second-order information without
approximating the full (inverse) Hessian matrix, which could be very large
for some networks.  Note that the use of multiple methods is not currently
supported.}

\item{itnmax, control, hessian}{Additional arguments passed to
\code{\link[optimx]{optimx}}}
}
\description{
Optimize a mistnet model using the \code{\link[optimx]{optimx}} package
}

