% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_distribution.R
\name{random_sample}
\alias{random_sample}
\title{Sample random numbers from a probability distribution}
\usage{
random_sample(distribution, ...)
}
\arguments{
\item{distribution}{an \code{\link{error_distribution}} object}

\item{...}{additional arguments passed to \code{\link{get_values}}}
}
\description{
Sample random numbers from a probability distribution
}

