% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{eNO}
\alias{deNO}
\alias{eNO}
\alias{reNO}
\title{Normal distribution with empirical mean and variance}
\usage{
eNO()

deNO(x, log, ...)

reNO(n, x, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{log}{logical; if TRUE, probabilities p are instead given as log(p)}

\item{n}{number of samples}

\item{mu}{vector of location parameters (not used)}
}
\description{
This distribution produces a Gaussian distribution whose moments are
dynamically set to match \code{mean(x)} and \code{sd(x)}. It is used for
empirical-Bayesian descriptions of the distribution of \code{x} (similar to
the distribution of coefficients in a mixed effects model).
}

