% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_values}
\alias{get_values}
\title{Get parameter values from a distribution object}
\usage{
get_values(distribution, ..., adjusted_values)
}
\arguments{
\item{distribution}{a distribution object}

\item{...}{additional arguments, which will override values contained within
the \code{distribution} object or in the \code{adjusted_values}}

\item{adjusted_values}{a \code{list} of adjusted values (for adjustable
parameters)}
}
\description{
Get parameter values from a distribution object
}

