% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{inflatable}
\alias{inflatable}
\alias{inflate}
\title{inflate an \code{inflatable} vector}
\usage{
inflatable(x, rep_rows, n)

inflate(x)
}
\arguments{
\item{x}{a vector}

\item{rep_rows}{logical. Should values be repeated across rows (TRUE) or across
columns (FALSE)?}

\item{n}{number of times to repeat each element}
}
\description{
description
}

