% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_distribution.R
\name{IU}
\alias{IU}
\alias{dIU}
\title{Improper uniform distribution}
\usage{
IU()

dIU(x, mu)
}
\description{
This distribution produces a flat prior with no bounds.  Its gradient and
log-density are always zero. It's improper because its integral is not one
(and is not finite).
}

