% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_density.R
\name{log_density}
\alias{log_density}
\title{Calculate the log-likelihood of a network object}
\usage{
log_density(network, state, par, include_penalties = FALSE, ...)
}
\arguments{
\item{network}{A \code{network} object}

\item{state}{An optional \code{network_state} object. If not included,
state will be calculated by \code{\link{feedforward}}.}

\item{par}{a vector of parameters}

\item{include_penalties}{Should penalty terms be included in the
log-likelihood?}

\item{...}{Additional arguments (currently not used.)}
}
\description{
Calculate the log-likelihood of a network object
}

