% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_distribution.R
\name{make_gamlss_distribution}
\alias{error_distribution}
\alias{make_gamlss_distribution}
\title{Make an \code{error_distribution} from a gamlss distribution}
\usage{
make_gamlss_distribution(abbreviation, ...)
}
\arguments{
\item{abbreviation}{An abbreviation matching a \link[gamlss.dist]{gamlss.family},
e.g. "NO" for the normal distribution or "ZIP" for the zero-inflated
Poisson distribution}

\item{...}{Additional arguments, possibly including \code{bd} (binomial denominator),
\code{sigma} (scale), \code{nu} (shape), or \code{tau} (shape),
depending on the distribution}
}
\value{
An \code{error_distribution} object, consisting of the following
   functions:
\itemize{
   \item{\code{log_density(x, mu)}: log probability of the distribution with
       location parameter \code{mu} evaluated at \code{x}.}
   \item{\code{sample(n, mu)}: draw \code{n} random samples from the distribution
       with location parameter \code{mu}.}
   \item{Functions for calculating partial derivatives evaluated at \code{x}
       with regard to mu (\code{dldm}), sigma (\code{dldd}), nu
       (\code{dldv}), and tau (\code{dldt}). For a few distributions,
       \code{dldx}} is also defined: these distributions can be used as
       prior distributions on the model's weights.
}
}
\description{
Make an \code{error_distribution} from a gamlss distribution
}
\examples{
distribution = make_gamlss_distribution("NO", sigma = 1/3)

# Sample 10 random values with mean=2 (and sigma=1/3 as defined above)
samples = distribution$sample(n = 10, mu = 2)

# find the log_density of those samples under a distribution with mean=1
distribution$log_density(samples, mu = 1)

# The gradient of the log_density with respect to the mean indicates that
# the log_density would be higher if mu were larger (more positive)
distribution$dldm(samples, mu = 1)

# The gradient with respect to x shows that the log_density would also be
# higher if we reduced the values of x (more negative)
distribution$dldx(samples, mu = 1)

# The gradient with respect to x isn't defined for all distributions,
however. For example, this code would return an error that
" gradient with respect to x (dldx) is not defined for the distribution PO"
\dontrun{
another_distribution = make_gamlss_distribution("PO")
another_distribution$dldx(x = 1:10, mu = 1)
}
}

