% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistnet.R
\name{mistnet}
\alias{mistnet}
\title{Build and fit a neural network with random effects}
\usage{
mistnet(x, y, n_z, activators, n_hidden, error_distribution, priors,
  fit = TRUE, starttests = FALSE, sigma = NULL, tau = NULL, nu = NULL,
  bd = NULL)
}
\arguments{
\item{x}{A numeric matrix of predictor variables}

\item{y}{A numeric or integer matrix of response variables}

\item{n_z, }{The number of latent random variables to include as predictors
alongside x}

\item{activators}{A list of \code{\link{activator}} objects, one per
network layer}

\item{n_hidden}{An integer vector determining the number of hidden nodes in
each hidden layer. Its length should be one less than that of the
\code{activators} list.}

\item{error_distribution}{Either an \code{\link{error_distribution}} object
or a \link[gamlss.dist]{gamlss.family} abbreviation (such as "NO" for the
normal distribution)}

\item{priors}{[[Add me]]}

\item{fit}{Logical. Should the model be fitted or should an untrained model
be returned. Defaults to TRUE}

\item{starttests}{Should \code{\link[optimx]{optimx}}'s \code{starttests} be
run? Can be useful for identifying errors but is not usually needed.}

\item{sigma}{(optional) scale parameter passed to
\code{\link{make_gamlss_distribution}}}

\item{tau}{(optional) shape parameter passed to
\code{\link{make_gamlss_distribution}}}

\item{nu}{(optional) shape parameter passed to
\code{\link{make_gamlss_distribution}}}

\item{bd}{(optional) binomial denominator passed to
\code{\link{make_gamlss_distribution}}}
}
\value{
A \code{network} object
}
\description{
Build and fit a neural network with random effects
}

