% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistnet.R
\name{mistnet}
\alias{mistnet}
\title{Build and fit a neural network with random effects}
\usage{
mistnet()
}
\value{
A \code{network} object
}
\description{
Build and fit a neural network with random effects
}

