% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.network}
\alias{predict.mistnet_network}
\alias{predict.network}
\title{Make predictions from a trained network}
\usage{
\method{predict}{network}(object, newdata, ...)

\method{predict}{mistnet_network}(object, newdata, n_samples, ...)
}
\arguments{
\item{object}{an object of class \code{network} or \code{mistnet_network}}

\item{newdata}{a matrix with the same columns as \code{x}}

\item{...}{further arguments passed to or from other methods.}

\item{n_samples}{(for \code{mistnet_networks} only). The number of Monte
Carlo samples to use}
}
\value{
For generic networks, a \code{network_state} object, as in
   \code{\link{feedforward}}. For mistnet networks, a list of such objects,
   a list of such object (one for each Monte Carlo sample)
}
\description{
Predictions are made by feeding new data through the network (see
\code{\link{feedforward}}.  For networks with latent variables
(\code{mistnet_network}s), these variables are sampled randomly from their
prior distribution and then each batch of samples is fed through the network
}

